+++ examples/lib/stories/camera_and_viewport/zoom.dart
@@ -17,6 +17,8 @@ class ZoomGame extends BaseGame with ScrollDetector, ScaleDetector {
     final flameSprite = await loadSprite('flame.png');
 
     viewport = FixedResolutionViewport(viewportResolution);
+    camera.setRelativeOffset(Anchor.center);
+    camera.cameraSpeed = 1;
 
     final flameSize = Vector2(149, 211);
     add(
@@ -25,13 +27,12 @@ class ZoomGame extends BaseGame with ScrollDetector, ScaleDetector {
         size: flameSize,
       )..anchor = Anchor.center,
     );
-    camera.followComponent(flame);
   }
 
   static const zoomPerScrollUnit = 0.001;
   @override
-  void onScroll(PointerScrollInfo event) {
-    camera.zoom += event.scrollDelta.game.y * zoomPerScrollUnit;
+  void onScroll(PointerScrollInfo info) {
+    camera.zoom += info.scrollDelta.game.y * zoomPerScrollUnit;
   }
 
   @override