+++ packages/flame/lib/src/game/game_render_box.dart
@@ -13,15 +13,14 @@ import 'game_loop.dart';
 class GameRenderBox extends RenderBox with WidgetsBindingObserver {
   BuildContext buildContext;
   Game game;
-  late GameLoop gameLoop;
+  GameLoop? gameLoop;
 
   GameRenderBox(this.buildContext, this.game) {
-    gameLoop = GameLoop(gameLoopCallback);
     WidgetsBinding.instance!.addTimingsCallback(game.onTimingsCallback);
   }
 
   @override
-  bool get sizedByParent => true;
+  bool get isRepaintBoundary => true;
 
   @override
   void performResize() {
@@ -33,6 +32,8 @@ class GameRenderBox extends RenderBox with WidgetsBindingObserver {
   void attach(PipelineOwner owner) {
     super.attach(owner);
     game.attach(owner, this);
+
+    final gameLoop = this.gameLoop = GameLoop(gameLoopCallback);
 
     game.pauseEngineFn = gameLoop.pause;
     game.resumeEngineFn = gameLoop.resume;
@@ -48,7 +49,8 @@ class GameRenderBox extends RenderBox with WidgetsBindingObserver {
   void detach() {
     super.detach();
     game.detach();
-    gameLoop.stop();
+    gameLoop?.dispose();
+    gameLoop = null;
     _unbindLifecycleListener();
   }
 
@@ -58,6 +60,11 @@ class GameRenderBox extends RenderBox with WidgetsBindingObserver {
     }
     game.update(dt);
     markNeedsPaint();
+  }
+
+  @override
+  void performLayout() {
+    size = constraints.biggest;
   }
 
   @override
@@ -81,6 +88,6 @@ class GameRenderBox extends RenderBox with WidgetsBindingObserver {
     game.lifecycleStateChange(state);
   }
 
-  // ignore: annotate_overrides
+  @override
   Size computeDryLayout(BoxConstraints constraints) => constraints.biggest;
 }